week1/flask-mysql-repl/app/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
